Deprecated validations
moved to alpha_compatibility
