--- lib/pavlov/validation_error.rb
@@ -1,4 +1,0 @@
-module Pavlov
-  class ValidationError < StandardError
-  end
-end
